Implement caching for import_* calls if context isn't forced.
Basically the notion is for `import_*` tags to cache the content
rendered if it's just using the global render context; that cache lives
for the duration of the minion render (or state render if triggered via
minion side).

This is an optimization for heavy users of `import_yaml`; in general
pillarstack is a better approach if one can structure their
configurables that way, but for codebases not yet there (or limited due
to other reasons) improving `import_*` performance is beneficial.

For example, the codebase I work against- this reduces our pillar render time
from ~74s to ~65s; that's something of an edge case for our renders, but
the complexity of the target is why it's so large, and why an optimization
like this has value.

This logic is wired such that if context is enabled for the
`import`--meaning the target does not have a stable render context (like
one has in a state or while doing a pillar render)--caching is disabled
automatically.